loki/requirements.txt
# Loki AI Desktop Assistant — Requirements

# Core
openai>=1.12.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Audio - Input
sounddevice>=0.4.6
webrtcvad-wheels>=2.0.10
openai-whisper>=20231117

# Audio - Output (edge-tts is primary, pyttsx3 fallback)
edge-tts>=6.1.9
pyttsx3>=2.90
soundfile>=0.12.1
pygame>=2.5.2
playsound>=1.3.0

# Wakeword (optional, falls back to whisper)
# pvporcupine>=3.0.0

# GUI
PyQt6>=6.6.0

# System Control (Windows)
pycaw>=20231129
screen-brightness-control>=0.22.0
pyautogui>=0.9.54
psutil>=5.9.8
pywin32>=306

# Clipboard
pyperclip>=1.8.2

# Web features
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# PDF
PyMuPDF>=1.23.0

# Git
gitpython>=3.1.40

# Security / Vault
cryptography>=41.0.0

# File operations
pathlib2>=2.3.7; python_version < '3.6'
watchdog>=3.0.0

# ML
numpy>=1.24.0
torch>=2.1.0

# Testing
pytest>=7.4.3
pytest-qt>=4.3.1
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
